Move addressable/target_addressable to build_graph.
These are internal API, they shouldn't need deprecating.
